share build

+++ app/src/main/java/com/mohit/installer/OnAttachmentDownloadListener.java
@@ -4,5 +4,5 @@ public interface OnAttachmentDownloadListener {
     void onAttachmentDownloadedSuccess();
     void onAttachmentDownloadedError();
     void onAttachmentDownloadedFinished();
-    void onAttachmentDownloadUpdate(int percent);
+    void onAttachmentDownloadUpdate(float percent);
 }
